[ALPS06729315] vdec: support vpud 64bit[1/2]
vdec: support vpud 64bit

Test: build pass

MTK-Commit-Id: 46dae9ecab129087740e915b0189d11031d3f95e

CR-Id: ALPS06729315
Feature: Video Driver

+++ drivers/media/platform/mtk-vcodec/vdec_vcu_if.h
@@ -31,7 +31,7 @@ struct vdec_vcu_inst {
 	enum ipi_id id;
 	void *vsi;
 	int32_t failure;
-	uint32_t inst_addr;
+	uint64_t inst_addr;
 	unsigned int signaled;
 	unsigned int signaled_res;
 	struct mtk_vcodec_ctx *ctx;
